requirements.txt
streamlit>=1.30.0
pyautogui>=0.9.54
requests>=2.31.0
Pillow>=10.0.0
pydantic>=2.0.0
